backend/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
black==21.9b0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
click==8.0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.5
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.0
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.0.2
regex==2021.10.23
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zope.interface==5.4.0
